firmware/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abyss-firmware"
version = "0.1.0"
description = "Diving App for Raspberry"
authors = [{name = "Mathéo Lopez, Nathan Füllemann, Arno Tribolet", email = "*@hes-so.ch"}]
dependencies = [
    "st7789>=1.0.0",
    "Pillow",
    "numpy",
    "RPi.GPIO",
    "spidev"]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = ["pytest>=6.0", "black", "flake8"]

[project.scripts]
hello-world = "hello_world.main:main"

[tool.setuptools.packages.find]
where = ["src"]
